+++ src/main/java/mcpecommander/mobultion/entity/entities/spiders/EntitySpeedySpider.java
@@ -49,7 +49,7 @@ public class EntitySpeedySpider extends EntityAnimatedSpider{
     @Override
     public double getMountedYOffset()
     {
-        return (double)(this.height * 0.6F);
+        return this.height * 0.6F;
     }
 
     @Override
