+++ src/main/java/mcpecommander/mobultion/entity/entities/spiders/EntityWitherSpider.java
@@ -53,7 +53,7 @@ public class EntityWitherSpider extends EntityAnimatedSpider {
 
 	@Override
 	public double getMountedYOffset() {
-		return (double) (this.height * 0.6F);
+		return this.height * 0.6F;
 	}
 	
 	@Override
