+++ src/main/java/mcpecommander/mobultion/entity/renderer/skeletonsRenderer/RenderMagmaSkeleton.java
@@ -101,10 +101,10 @@ public class RenderMagmaSkeleton extends RenderLiving<EntityMagmaSkeleton>{
             }
             else
             {
-                float f1 = (float)(i >> 24 & 255) / 255.0F;
-                float f2 = (float)(i >> 16 & 255) / 255.0F;
-                float f3 = (float)(i >> 8 & 255) / 255.0F;
-                float f4 = (float)(i & 255) / 255.0F;
+                float f1 = (i >> 24 & 255) / 255.0F;
+                float f2 = (i >> 16 & 255) / 255.0F;
+                float f3 = (i >> 8 & 255) / 255.0F;
+                float f4 = (i & 255) / 255.0F;
                 this.brightnessBuffer.put(f2);
                 this.brightnessBuffer.put(f3);
                 this.brightnessBuffer.put(f4);
