+++ src/main/java/mcpecommander/mobultion/items/ItemCorruptedBonemeal.java
@@ -46,10 +46,10 @@ public class ItemCorruptedBonemeal extends Item {
 						double d1 = itemRand.nextGaussian() * 0.02D;
 						double d2 = itemRand.nextGaussian() * 0.02D;
 						world.spawnParticle(EnumParticleTypes.VILLAGER_HAPPY,
-								(double) ((float) pos.getX() + itemRand.nextFloat()),
-								(double) pos.getY() + (double) itemRand.nextFloat()
+								pos.getX() + itemRand.nextFloat(),
+								pos.getY() + itemRand.nextFloat()
 										* worldIn.getBlockState(pos).getBoundingBox(worldIn, pos).maxY,
-								(double) ((float) pos.getZ() + itemRand.nextFloat()), 1, d0, d1, d2, 0.01d);
+								pos.getZ() + itemRand.nextFloat(), 1, d0, d1, d2, 0.01d);
 					}
 
 					return EnumActionResult.SUCCESS;
@@ -61,10 +61,10 @@ public class ItemCorruptedBonemeal extends Item {
 							double d1 = itemRand.nextGaussian() * 0.02D;
 							double d2 = itemRand.nextGaussian() * 0.02D;
 							world.spawnParticle(EnumParticleTypes.SPIT,
-									(double) ((float) pos.getX() + itemRand.nextFloat()),
-									(double) pos.getY() + (double) itemRand.nextFloat()
+									pos.getX() + itemRand.nextFloat(),
+									pos.getY() + itemRand.nextFloat()
 											* worldIn.getBlockState(pos).getBoundingBox(worldIn, pos).maxY,
-									(double) ((float) pos.getZ() + itemRand.nextFloat()), 1, d0, d1, d2, 0.01d);
+									pos.getZ() + itemRand.nextFloat(), 1, d0, d1, d2, 0.01d);
 						}
 						worldIn.setBlockState(pos, Blocks.AIR.getDefaultState());
 						itemstack.shrink(1);
