+++ src/main/java/mcpecommander/mobultion/potion/PotionVomit.java
@@ -36,7 +36,7 @@ public class PotionVomit extends Potion{
 				&& entityLivingBaseIn.isPotionActive(ModPotions.potionVomit)) {
 			
 			if(!isReady){
-				temp = 0.98f / (float)entityLivingBaseIn.getActivePotionEffect(ModPotions.potionVomit).getDuration();
+				temp = 0.98f / entityLivingBaseIn.getActivePotionEffect(ModPotions.potionVomit).getDuration();
 			}
 			this.isReady = true;
 			this.mul += temp/2.5f;
@@ -58,9 +58,9 @@ public class PotionVomit extends Potion{
         Tessellator tessellator = Tessellator.getInstance();
         BufferBuilder bufferbuilder = tessellator.getBuffer();
         bufferbuilder.begin(7, DefaultVertexFormats.POSITION_TEX);
-        bufferbuilder.pos(0.0D, (double)e.getResolution().getScaledHeight(), -90.0D).tex(0.0D, 1.0D).endVertex();
-        bufferbuilder.pos((double)e.getResolution().getScaledWidth(), (double)e.getResolution().getScaledHeight(), -90.0D).tex(1.0D, 1.0D).endVertex();
-        bufferbuilder.pos((double)e.getResolution().getScaledWidth(), 0.0D, -90.0D).tex(1.0D, 0.0D).endVertex();
+        bufferbuilder.pos(0.0D, e.getResolution().getScaledHeight(), -90.0D).tex(0.0D, 1.0D).endVertex();
+        bufferbuilder.pos(e.getResolution().getScaledWidth(), e.getResolution().getScaledHeight(), -90.0D).tex(1.0D, 1.0D).endVertex();
+        bufferbuilder.pos(e.getResolution().getScaledWidth(), 0.0D, -90.0D).tex(1.0D, 0.0D).endVertex();
         bufferbuilder.pos(0.0D, 0.0D, -90.0D).tex(0.0D, 0.0D).endVertex();
         tessellator.draw();
         GlStateManager.depthMask(true);
